# laravel-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API – Aplikacija za izložbe fotografija",
        "description": "Swagger dokumentacija Laravel API-ja",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Lokalno (Docker)"
        }
    ],
    "paths": {
...
        "/api/izlozbe": {
            "get": {
                "tags": [
                    "Izložbe"
                ],
                "summary": "Lista izložbi (sa filterima/paginacijom)",
                "operationId": "a8deab27d9f703cac2820dc4d16aad8e",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Izložbe"
                ],
                "summary": "Kreiranje izložbe (admin)",
                "operationId": "8dbc4fc1345e0f2b1bd0410f783246ef",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "naziv",
                                    "datum",
                                    "lokacija"
                                ],
                                "properties": {
                                    "naziv": {
                                        "type": "string",
                                        "example": "Izložba 2026"
                                    },
                                    "datum": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-03-10"
                                    },
                                    "lokacija": {
                                        "type": "string",
                                        "example": "Beograd"
                                    },
                                    "opis": {
                                        "type": "string",
                                        "example": "Opis..."
                                    },
                                    "dostupna_mesta": {
                                        "type": "integer",
                                        "example": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Kreirano"
                    },
                    "403": {
                        "description": "Zabranjeno"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/izlozbe/{id}": {
            "get": {
                "tags": [
                    "Izložbe"
                ],
                "summary": "Detalji izložbe",
                "operationId": "eacbe74b71ad07f1d64c52393698253d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Nije pronađena"
                    }
                }
            },
            "put": {
                "tags": [
                    "Izložbe"
                ],
                "summary": "Izmena izložbe (admin)",
                "operationId": "1e8ef8211e309bf261a52d9a1f850386",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "naziv",
                                    "datum",
                                    "lokacija"
                                ],
                                "properties": {
                                    "naziv": {
                                        "type": "string",
                                        "example": "Izložba 2026"
                                    },
                                    "datum": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-03-10"
                                    },
                                    "lokacija": {
                                        "type": "string",
                                        "example": "Beograd"
                                    },
                                    "opis": {
                                        "type": "string",
                                        "example": "Opis..."
                                    },
                                    "dostupna_mesta": {
                                        "type": "integer",
                                        "example": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Nije pronađena"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Izložbe"
                ],
                "summary": "Brisanje izložbe (admin)",
                "operationId": "a779745ec1e2296f89eaf870ef0312de",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Obrisano"
                    },
                    "404": {
                        "description": "Nije pronađena"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/izlozbe/{id}/fotografije": {
            "get": {
                "tags": [
                    "Izložbe"
                ],
                "summary": "Fotografije za izložbu",
                "operationId": "1bcb3e17f8964fdbd62364b30765ae09",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Nije pronađena"
                    }
                }
            }
        },
        "/api/izlozbe/{id}/prijave": {
            "get": {
                "tags": [
                    "Izložbe"
                ],
                "summary": "Prijave za izložbu (sa korisnikom)",
                "operationId": "6e30109b674f4bcd25d871e24e601d54",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Nije pronađena"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Unesi token u formatu: Bearer {token}",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}